[#21] Fixing auth problems
- request all from home view (with token)
- request real podcast from podcast list view
