+++ src/podcast/ViewModels/PodcastViewModel.swift
@@ -1,11 +1,10 @@
 import Foundation
 
 
-
 class PodcastListViewModel: ObservableObject {
-    var podcasts: [PodcastItem] = []
-    var episodes: [Episode] = []
-//    
+    @Published var podcasts: [PodcastItem] = []
+    @Published var episodes: [Episode] = []
+
     init() {
         self.getPodcasts()
         self.getAllEpisodes(limit: 5)
@@ -22,13 +21,10 @@ class PodcastListViewModel: ObservableObject {
         WebService().getPodcasts(token: token){ result in
             switch result{
                 case .success(let podcasts):
-                    self.podcasts = podcasts
-                    print("Found podcasts \(self.podcasts)")
-//                    DispatchQueue.main.async {
-//                        self.podcasts = podcasts
-//                        print("Found podcasts \(self.podcasts)")
-//
-//                    }
+                    DispatchQueue.main.async {
+                        self.podcasts = podcasts
+                        print("Found podcasts \(self.podcasts)")
+                    }
                 case .failure(let error):
                     print("API problems: \(error.localizedDescription)")
             }
